requirements.txt
markdown>=3.4.1
pyyaml>=6.0
jinja2>=3.1.2
python-frontmatter>=1.0.0
requests>=2.28.1
beautifulsoup4>=4.11.1
pygments>=2.13.0
django>=4.2.0 